Fixes and utils:
- Added /clear to clear your inventory.
- Added building utils menu with access to debug stick (this menu will have more features in the future), additionally added /debugstick
- Added basic tab menu to show name and primary role.
- Fixed bug where name wasn't displayed correctly when leaving an owned plot.
- Fixed a bug with the region requests not working in the staff menu.
- Added command feedback for /plotsystem selectiontool
- Added /spawn and reworked the spawn button in the navigation menu to teleport to the centre of spawn.
- Protected the Lobby and set gamemode to adventure.
- When falling from the Lobby you get teleported back to the spawnpoint.
- Fixed globalSQL.getDouble, it was returning an int.

+++ src/main/java/me/bteuk/network/sql/GlobalSQL.java
@@ -83,7 +83,7 @@ public class GlobalSQL {
 
             if (results.next()) {
 
-                return results.getInt(1);
+                return results.getDouble(1);
 
             } else {
